requirements.txt
torch
torchvision
transformers
easyocr
opencv-python-headless
Pillow
streamlit
matplotlib
pandas==1.5.3
numpy==1.23.5
pytest